// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  CreateCard: typeof CreateCard;
  CreateInterviewCard: typeof CreateInterviewCard;
  DailyInteractions: typeof DailyInteractions;
  FindInterview: typeof FindInterview;
  FindTest: typeof FindTest;
  GeneratedCardSearch: typeof GeneratedCardSearch;
  GeneratedInterviewCardSearch: typeof GeneratedInterviewCardSearch;
  GetCard: typeof GetCard;
  GetHistory: typeof GetHistory;
  GetInterview: typeof GetInterview;
  GetInterviewAnswer: typeof GetInterviewAnswer;
  GetInterviewTags: typeof GetInterviewTags;
  GetTags: typeof GetTags;
  GetTest: typeof GetTest;
  GetUserAnswer: typeof GetUserAnswer;
  InterviewAnalytics: typeof InterviewAnalytics;
  InterviewWishlist: typeof InterviewWishlist;
  LeaderBoard: typeof LeaderBoard;
  LikeCount: typeof LikeCount;
  PushHeatMap: typeof PushHeatMap;
  TotalInteractions: typeof TotalInteractions;
  UserLike: typeof UserLike;
  createEmbedding: typeof createEmbedding;
  create_interview: typeof create_interview;
  create_mcq: typeof create_mcq;
  fetchFiles: typeof fetchFiles;
  getAllFeaturedCards: typeof getAllFeaturedCards;
  getAllFeaturedInterviewCards: typeof getAllFeaturedInterviewCards;
  getAllGeneratedCards: typeof getAllGeneratedCards;
  getAll_interviewGeneratedCards: typeof getAll_interviewGeneratedCards;
  messages: typeof messages;
  openai: typeof openai;
  pushAnswer: typeof pushAnswer;
  pushInterviewAnswer: typeof pushInterviewAnswer;
  uploadPdF: typeof uploadPdF;
  wishlist: typeof wishlist;
}>;
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;